+++ src/main/java/com/flightstats/hub/dao/CommonContentService.java
@@ -1,17 +1,14 @@
 package com.flightstats.hub.dao;
 
 import com.diffplug.common.base.Errors;
-import com.flightstats.hub.app.HubProperties;
-import com.flightstats.hub.app.HubServices;
+import com.flightstats.hub.app.InFlightService;
 import com.flightstats.hub.dao.aws.MultiPartParser;
 import com.flightstats.hub.exception.ContentTooLargeException;
 import com.flightstats.hub.metrics.ActiveTraces;
 import com.flightstats.hub.metrics.Traces;
 import com.flightstats.hub.model.*;
 import com.flightstats.hub.time.TimeService;
-import com.flightstats.hub.util.Sleeper;
 import com.google.common.base.Optional;
-import com.google.common.util.concurrent.AbstractIdleService;
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 import org.slf4j.Logger;
@@ -19,18 +16,13 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Collection;
 import java.util.SortedSet;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 
 /**
  * Performs common logic for all ContentServices.
  */
 public class CommonContentService implements ContentService {
     private final static Logger logger = LoggerFactory.getLogger(CommonContentService.class);
-
-    private final Integer shutdown_wait_seconds = HubProperties.getProperty("app.shutdown_wait_seconds", 10);
-    private final AtomicInteger inFlight = new AtomicInteger();
 
     @Inject
     private TimeService timeService;
@@ -39,22 +31,12 @@ public class CommonContentService implements ContentService {
     @Named(ContentService.IMPL)
     private ContentService contentService;
 
-    public CommonContentService() {
-        HubServices.registerPreStop(new CommonContentServiceShutdown());
-    }
-
-    private <X> X inFlight(Supplier<X> supplier) {
-        try {
-            inFlight.incrementAndGet();
-            return supplier.get();
-        } finally {
-            inFlight.decrementAndGet();
-        }
-    }
+    @Inject
+    private InFlightService inFlightService;
 
     @Override
     public ContentKey insert(String channelName, Content content) throws Exception {
-        return inFlight(Errors.rethrow().wrap(() -> doInsert(channelName, content)));
+        return inFlightService.inFlight(Errors.rethrow().wrap(() -> doInsert(channelName, content)));
     }
 
     private ContentKey doInsert(String channelName, Content content) throws Exception {
@@ -80,7 +62,7 @@ public class CommonContentService implements ContentService {
 
     @Override
     public Collection<ContentKey> insert(BulkContent bulkContent) throws Exception {
-        return inFlight(Errors.rethrow().wrap(() -> {
+        return inFlightService.inFlight(Errors.rethrow().wrap(() -> {
             MultiPartParser multiPartParser = new MultiPartParser(bulkContent);
             multiPartParser.parse();
             return contentService.insert(bulkContent);
@@ -89,7 +71,7 @@ public class CommonContentService implements ContentService {
 
     @Override
     public boolean historicalInsert(String channelName, Content content) throws Exception {
-        return inFlight(Errors.rethrow().wrap(() -> {
+        return inFlightService.inFlight(Errors.rethrow().wrap(() -> {
             content.packageStream();
             return contentService.historicalInsert(channelName, content);
         }));
@@ -139,32 +121,5 @@ public class CommonContentService implements ContentService {
     public void notify(ChannelConfig newConfig, ChannelConfig oldConfig) {
         contentService.notify(newConfig, oldConfig);
     }
-
-    private void waitForInFlight() {
-        logger.info("waiting for in-flight to complete " + inFlight.get());
-        long start = System.currentTimeMillis();
-        while (inFlight.get() > 0) {
-            logger.info("still waiting for in-flight to complete " + inFlight.get());
-            Sleeper.sleep(1000);
-            if (System.currentTimeMillis() > (start + shutdown_wait_seconds * 1000)) {
-                break;
-            }
-        }
-        logger.info("completed waiting for in-flight to complete " + inFlight.get());
-    }
-
-
-    private class CommonContentServiceShutdown extends AbstractIdleService {
-        @Override
-        protected void startUp() throws Exception {
-            //do nothing
-        }
-
-        @Override
-        protected void shutDown() throws Exception {
-            waitForInFlight();
-        }
-    }
-
 
 }
